Move general methods from PoTokenWebView to PoTokenGenerator
